+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/fivetran/go-fivetran/compare/v0.5.5...HEAD)
+## [Unreleased](https://github.com/fivetran/go-fivetran/compare/v0.5.6...HEAD)
+
+## [0.5.6](https://github.com/fivetran/go-fivetran/compare/v0.5.5...v0.5.6) - 2022-04-26
+
+## Fixed
+- `ConnectorConfig.APIKeys` field transformed to type `[]string`
+- `ConnectorConfig.AccountIds` field transfromed to type `[]string`
 
 ## [0.5.5](https://github.com/fivetran/go-fivetran/compare/v0.5.4...v0.5.5) - 2022-04-20